Move player Fix some crash bugs
