+++ furusiya/model/party/player.py
